serval-dfo: apply resolved spec decisions D1-D4
D1: Value.int widened to i128 — the dynamic-path Overflow loss point for
big u64 is gone (binary decodeValue passes Header.int through; json
parses i128; issue actual/expected payloads simplify). D2: string→float
coercion restricted to finite decimal — 'inf'/'nan' spellings rejected.
D3: digit separators rejected in string→int/float coercion — wire data
is not Zig source. D4: non-finite float→int stays Overflow. Pinning
tests updated to the resolved semantics; SPEC.md §10 records sign-off.

+++ src/core/value.zig
@@ -6,7 +6,9 @@
 pub const Value = union(enum) {
     null,
     bool: bool,
-    int: i64,
+    // serval-dfo (D1): i128 so the full u64+i64 range survives dynamic
+    // paths losslessly.
+    int: i128,
     float: f64,
     string: []const u8,
     bytes: []const u8,
